salesforce_connection/requirements.txt
Office365-REST-Python-Client
python-dotenv
